[ etc ] 타입 분리
