+++ src/type/editSchedule/editDateType.ts
@@ -1,0 +1,6 @@
+export interface editDateStateTypes {
+  year: number;
+  month: number;
+  date: number;
+  dayOfWeek: number;
+}
