+++ src/type/editSchedule/editScheduleType.ts
@@ -1,0 +1,8 @@
+export interface editScheduleType {
+    idx: number;
+    studentName: string;
+    subject: string;
+    startTime: string;
+    endTime: string;
+  }
+
